% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_sentiment.R
\name{highlight_sentiment}
\alias{highlight_sentiment}
\title{Highlight Sentiment}
\usage{
highlight_sentiment(x)
}
\arguments{
\item{x}{A \code{sentimentr::sentiment_by} object}
}
\value{
HTML for use in shiny app
}
\description{
Highlight Sentiment
}
\examples{
\dontrun{ highlight_sentiment(x)}
}
