% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_sentiment.R
\name{highlight_sentimentUI}
\alias{highlight_sentimentUI}
\title{Highlight Sentiment UI}
\usage{
highlight_sentimentUI(rmd = FALSE)
}
\value{
Style Tags
}
\description{
Highlight Sentiment UI
}
\examples{
\dontrun{ highlight_sentimentUI()}
}
